refactor: 상수를 하위 파일로 분리
