+++ src/js/utils/constants/dom.js
@@ -1,25 +1,3 @@
-export const LOTTO_SETTINGS = {
-  LOTTO_NUMBER_SIZE: 6,
-  BONUS_NUMBER_SIZE: 1,
-
-  MIN_LOTTO_NUMBER: 1,
-  MAX_LOTTO_NUMBER: 45,
-  LOTTO_PRICE: 1000,
-}
-
-export const UI_SETTINGS = {
-  DEFAULT_VISIBILITY: false,
-}
-
-export const ALERT_MESSAGES = {
-  UNDER_MIN_PRICE: `최소 ${LOTTO_SETTINGS.LOTTO_PRICE}원 이상의 금액을 입력해야 합니다.`,
-  NOT_INTEGER_PRICE: `구입 금액은 정수로 입력해야합니다.`,
-
-  EMPTY_RESULT_INPUT: '당첨 번호와 보너스 번호를 입력해주세요.',
-  DUPLICATED_NUMBERS_EXIST: '로또 번호에 중복이 있습니다.',
-  NUMBERS_OUT_OF_RANGE: '로또 번호는 1 ~ 45 사이의 숫자여야 합니다.',
-}
-
 export const DOM_CLASSES = {
   MONEY_INPUT_CONTAINER: 'money-input-div',
   MONEY_FORM: 'money-input-form',
@@ -46,12 +24,4 @@ export const DOM_CLASSES = {
 
 export const DOM_IDS = {
   APP: 'app',
-}
-
-export const PRIZE = {
-  FIRST: 2000000000,
-  SECOND: 30000000,
-  THIRD: 1500000,
-  FOURTH: 50000,
-  FIFTH: 5000,
 }
