+++ src/js/utils/constants/settings.js
@@ -1,0 +1,20 @@
+export const LOTTO_SETTINGS = {
+  LOTTO_NUMBER_SIZE: 6,
+  BONUS_NUMBER_SIZE: 1,
+
+  MIN_LOTTO_NUMBER: 1,
+  MAX_LOTTO_NUMBER: 45,
+  LOTTO_PRICE: 1000,
+}
+
+export const UI_SETTINGS = {
+  DEFAULT_VISIBILITY: false,
+}
+
+export const PRIZE = {
+  FIRST: 2000000000,
+  SECOND: 30000000,
+  THIRD: 1500000,
+  FOURTH: 50000,
+  FIFTH: 5000,
+}
